Ignore some importlib-resources type errors
During static type checking, always install the old version of
importlib-resources that works with Python 3.6 because the contents()
and path() functions were removed in the latest version but the
alternative to them only works with newer Python versions.

Ignore types for files().joinpath() because this old version does not
provide type definitions for its return type.